Add AltGr to most layers
In layout variants that use AltGr for entering non ascii characters, it
is a very different key from Alt. For an example, see:

https://altgr-weur.eu/altgr-intl.html#ascii

This patch then adds AltGr just bellow alt in all layers except
pointer, where the spot was already taken.

The idea for the location was taken from miryoku.

While at it, I also changed the other home row keys to be identical in
both hands in the base layer.